+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/RobotConfigs.java
@@ -6,7 +6,10 @@ import com.pedropathing.ftc.localization.constants.PinpointConstants;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 import org.firstinspires.ftc.teamcode.subsystems.DriveSubsystem;
 import org.firstinspires.ftc.teamcode.subsystems.IntakeSubsystem;
-import org.firstinspires.ftc.teamcode.subsystems.LauncherSubsystem;
+import org.firstinspires.ftc.teamcode.subsystems.launcher.config.LauncherFeederConfig;
+import org.firstinspires.ftc.teamcode.subsystems.launcher.config.LauncherFlywheelConfig;
+import org.firstinspires.ftc.teamcode.subsystems.launcher.config.LauncherHoodConfig;
+import org.firstinspires.ftc.teamcode.subsystems.launcher.config.LauncherTimingConfig;
 
 /**
  * Centralized robot-specific configuration selector.
@@ -19,31 +22,31 @@ public class RobotConfigs {
 
     /**
      * Gets the active FeederConfig for the current robot.
-     * Returns the static instance from LauncherSubsystem so Dashboard edits are applied.
+     * Returns the static instance from LauncherFeederConfig so Dashboard edits are applied.
      * @return FeederConfig19429 or FeederConfig20245 based on robot name
      */
-    public static LauncherSubsystem.FeederConfig getFeederConfig() {
+    public static LauncherFeederConfig getFeederConfig() {
         String robotName = RobotState.getRobotName();
         if ("DECODE_19429".equals(robotName)) {
-            return LauncherSubsystem.feederConfig19429;
+            return LauncherFeederConfig.feederConfig19429;
         } else {
-            // Default to 19429 config if robot name is unknown or is 19429
-            return LauncherSubsystem.feederConfig20245;
+            // Default to 20245 config if robot name is unknown or is 20245
+            return LauncherFeederConfig.feederConfig20245;
         }
     }
 
     /**
      * Gets the active HoodConfig for the current robot.
-     * Returns the static instance from LauncherSubsystem so Dashboard edits are applied.
+     * Returns the static instance from LauncherHoodConfig so Dashboard edits are applied.
      * @return HoodConfig19429 or HoodConfig20245 based on robot name
      */
-    public static LauncherSubsystem.HoodConfig getHoodConfig() {
+    public static LauncherHoodConfig getHoodConfig() {
         String robotName = RobotState.getRobotName();
         if ("DECODE_19429".equals(robotName)) {
-            return LauncherSubsystem.hoodConfig19429;
+            return LauncherHoodConfig.hoodConfig19429;
         } else {
-            // Default to 19429 config if robot name is unknown or is 19429
-            return LauncherSubsystem.hoodConfig20245;
+            // Default to 20245 config if robot name is unknown or is 20245
+            return LauncherHoodConfig.hoodConfig20245;
         }
     }
 
@@ -112,29 +115,29 @@ public class RobotConfigs {
 
     /**
      * Gets the active Timing for the current robot.
-     * Returns the static instance from LauncherSubsystem so Dashboard edits are applied.
+     * Returns the static instance from LauncherTimingConfig so Dashboard edits are applied.
      * @return timing19429 or timing20245 based on robot name
      */
-    public static LauncherSubsystem.Timing getTiming() {
+    public static LauncherTimingConfig getTiming() {
         String robotName = RobotState.getRobotName();
         if ("DECODE_19429".equals(robotName)) {
-            return LauncherSubsystem.timing19429;
+            return LauncherTimingConfig.timing19429;
         } else {
-            return LauncherSubsystem.timing20245;
+            return LauncherTimingConfig.timing20245;
         }
     }
 
     /**
      * Gets the active FlywheelConfig for the current robot.
-     * Returns the static instance from LauncherSubsystem so Dashboard edits are applied.
+     * Returns the static instance from LauncherFlywheelConfig so Dashboard edits are applied.
      * @return flywheelConfig19429 or flywheelConfig20245 based on robot name
      */
-    public static LauncherSubsystem.FlywheelConfig getFlywheelConfig() {
+    public static LauncherFlywheelConfig getFlywheelConfig() {
         String robotName = RobotState.getRobotName();
         if ("DECODE_19429".equals(robotName)) {
-            return LauncherSubsystem.flywheelConfig19429;
+            return LauncherFlywheelConfig.flywheelConfig19429;
         } else {
-            return LauncherSubsystem.flywheelConfig20245;
+            return LauncherFlywheelConfig.flywheelConfig20245;
         }
     }
